+++ tests/python_tests/test_llm_pipeline.py
@@ -13,13 +13,13 @@ import torch
 
 from common import run_llm_pipeline_with_ref
 from ov_genai_test_utils import (
-    get_models_list,
     read_model,
     load_genai_pipe_with_configs,
     get_chat_models_list,
     model_tmp_path,
 )
 from utils.hugging_face import generation_config_to_hf
+from data.models import get_models_list
 
 #
 # e2e work
